# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FurniFlip API - AI-powered furniture photo editing and ad generation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  pro: 'pro',
} as const;

export interface User {
  id: number;
  email: string;
  plan: UserPlan;
  trialUsed: number;
  trialLimit: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface Project {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  originalImageUrl: string;
  imageCount: number;
  adCount: number;
  createdAt: string;
}

export type EditedImageType = typeof EditedImageType[keyof typeof EditedImageType];


export const EditedImageType = {
  background_removed: 'background_removed',
  staged: 'staged',
} as const;

export interface EditedImage {
  id: number;
  projectId: number;
  type: EditedImageType;
  imageUrl: string;
  /** @nullable */
  roomStyle?: string | null;
  createdAt: string;
}

export interface Ad {
  id: number;
  projectId: number;
  title: string;
  description: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  condition?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  originalImageUrl: string;
  images: EditedImage[];
  ads: Ad[];
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface ProjectStats {
  totalProjects: number;
  totalImages: number;
  totalAds: number;
  recentProjects: Project[];
}

export type StageInputRoomStyle = typeof StageInputRoomStyle[keyof typeof StageInputRoomStyle];


export const StageInputRoomStyle = {
  modern: 'modern',
  rustic: 'rustic',
  bohemian: 'bohemian',
  minimalist: 'minimalist',
  industrial: 'industrial',
  traditional: 'traditional',
  coastal: 'coastal',
  mid_century: 'mid_century',
} as const;

export interface StageInput {
  roomStyle: StageInputRoomStyle;
  /** @nullable */
  additionalPrompt?: string | null;
  /**
     * ID of an edited image to use as source (e.g. background-removed). Uses original if null.
     * @nullable
     */
  sourceImageId?: number | null;
}

export type AdInputCondition = typeof AdInputCondition[keyof typeof AdInputCondition];


export const AdInputCondition = {
  like_new: 'like_new',
  good: 'good',
  fair: 'fair',
  poor: 'poor',
} as const;

export interface AdInput {
  /** @nullable */
  price?: number | null;
  condition?: AdInputCondition;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  additionalContext?: string | null;
  /**
     * ID of edited image to feature in the ad. Uses original if null.
     * @nullable
     */
  imageId?: number | null;
}

export type UsageInfoPlan = typeof UsageInfoPlan[keyof typeof UsageInfoPlan];


export const UsageInfoPlan = {
  free: 'free',
  pro: 'pro',
} as const;

export interface UsageInfo {
  plan: UsageInfoPlan;
  trialUsed: number;
  trialLimit: number;
  isTrialExpired: boolean;
  hasActiveSubscription: boolean;
}

export interface CheckoutSession {
  url: string;
}