# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FurniFlip API - AI-powered furniture photo editing and ad generation
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}/remove-background:
    post:
      operationId: removeBackground
      tags: [images]
      summary: Remove the background from a project image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Background-removed image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditedImage"
        "402":
          description: Trial limit reached or no active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/stage:
    post:
      operationId: stageRoom
      tags: [images]
      summary: Stage the furniture in a styled room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageInput"
      responses:
        "200":
          description: Staged image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditedImage"
        "402":
          description: Trial limit reached or no active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/images:
    get:
      operationId: listProjectImages
      tags: [images]
      summary: List all edited images for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Edited images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EditedImage"
  /projects/{id}/ads:
    get:
      operationId: listProjectAds
      tags: [ads]
      summary: List all ads for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: generateAd
      tags: [ads]
      summary: Generate a Facebook Marketplace ad for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Generated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "402":
          description: Trial limit reached or no active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        plan:
          type: string
          enum: [free, pro]
        trialUsed:
          type: integer
        trialLimit:
          type: integer
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - plan
        - trialUsed
        - trialLimit
        - createdAt
    AuthInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
...
    EditedImage:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [background_removed, staged]
        imageUrl:
          type: string
        roomStyle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - imageUrl
        - createdAt
    StageInput:
      type: object
      properties:
        roomStyle:
          type: string
          enum: [modern, rustic, bohemian, minimalist, industrial, traditional, coastal, mid_century]
        additionalPrompt:
          type: ["string", "null"]
        sourceImageId:
          type: ["integer", "null"]
          description: ID of an edited image to use as source (e.g. background-removed). Uses original if null.
      required:
        - roomStyle
    Ad:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        condition:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - description
        - createdAt
    AdInput:
      type: object
      properties:
        price:
          type: ["number", "null"]
        condition:
          type: string
          enum: [like_new, good, fair, poor]
        location:
          type: ["string", "null"]
        additionalContext:
          type: ["string", "null"]
        imageId:
          type: ["integer", "null"]
          description: ID of edited image to feature in the ad. Uses original if null.
      required: []
...