# lib/api-zod/src/generated/types/adInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FurniFlip API - AI-powered furniture photo editing and ad generation
 * OpenAPI spec version: 0.1.0
 */
import type { AdInputCondition } from './adInputCondition';

export interface AdInput {
  /** @nullable */
  price?: number | null;
  condition?: AdInputCondition;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  additionalContext?: string | null;
  /**
     * ID of edited image to feature in the ad. Uses original if null.
     * @nullable
     */
  imageId?: number | null;
}