# lib/api-zod/src/generated/types/editedImage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FurniFlip API - AI-powered furniture photo editing and ad generation
 * OpenAPI spec version: 0.1.0
 */
import type { EditedImageType } from './editedImageType';

export interface EditedImage {
  id: number;
  projectId: number;
  type: EditedImageType;
  imageUrl: string;
  /** @nullable */
  roomStyle?: string | null;
  createdAt: Date;
}