# lib/api-zod/src/generated/types/stageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FurniFlip API - AI-powered furniture photo editing and ad generation
 * OpenAPI spec version: 0.1.0
 */
import type { StageInputRoomStyle } from './stageInputRoomStyle';

export interface StageInput {
  roomStyle: StageInputRoomStyle;
  /** @nullable */
  additionalPrompt?: string | null;
  /**
     * ID of an edited image to use as source (e.g. background-removed). Uses original if null.
     * @nullable
     */
  sourceImageId?: number | null;
}