# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FurniFlip API - AI-powered furniture photo editing and ad generation
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';

export interface User {
  id: number;
  email: string;
  plan: UserPlan;
  trialUsed: number;
  trialLimit: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: Date;
}